+++ src/lib.rs
@@ -163,6 +163,8 @@ mod repository;
 #[cfg(feature = "monitoring")]
 mod attributes;
 #[cfg(feature = "monitoring")]
+mod backoff;
+#[cfg(feature = "monitoring")]
 mod collector;
 #[cfg(feature = "monitoring")]
 mod monitor;